+++ src/lubrication/src/4C_lubrication_timint_implicit.hpp
@@ -11,8 +11,8 @@
 #include "4C_config.hpp"
 
 #include "4C_fem_discretization.hpp"
-#include "4C_inpar_lubrication.hpp"
 #include "4C_linalg_serialdensevector.hpp"
+#include "4C_lubrication_input.hpp"
 #include "4C_utils_parameter_list.fwd.hpp"
 
 #include <Epetra_Map.h>
